+++ src/main/kotlin/com/coder/gateway/models/CoderWorkspacesWizardModel.kt
@@ -1,0 +1,3 @@
+package com.coder.gateway.models
+
+data class CoderWorkspacesWizardModel(var loginModel: LoginModel, var workspaces: List<Workspace>)